feat: implement MagnifierView for precise cropping and add comprehensive documentation to core data models
