+++ android/src/main/kotlin/biz/cunning/cunning_document_scanner/fallback/DocumentScannerFileProvider.kt
@@ -2,5 +2,6 @@ package biz.cunning.cunning_document_scanner.fallback
 
 import androidx.core.content.FileProvider
 
+/// Custom FileProvider class to handle secure content:// Uri generation for image files.
 class DocumentScannerFileProvider: FileProvider() {
 }
