+++ android/src/main/kotlin/biz/cunning/cunning_document_scanner/fallback/models/Document.kt
@@ -1,15 +1,11 @@
 package biz.cunning.cunning_document_scanner.fallback.models
 
-/**
- * This class contains the original document photo, and a cropper. The user can drag the corners
- * to make adjustments to the detected corners.
- *
- * @param originalPhotoFilePath the photo file path before cropping
- * @param originalPhotoWidth the original photo width
- * @param originalPhotoHeight the original photo height
- * @param corners the document's 4 corner points
- * @constructor creates a document
- */
+/// Represents a scanned document, holding the file path, size dimensions, and cropping corner coordinates.
+///
+/// @param originalPhotoFilePath The local file path to the original un-cropped photo.
+/// @param originalPhotoWidth The width of the original photo in pixels.
+/// @param originalPhotoHeight The height of the original photo in pixels.
+/// @param corners The Quad representation of the four crop corner coordinates.
 class Document(
     val originalPhotoFilePath: String,
     private val originalPhotoWidth: Int,
