+++ android/src/main/kotlin/biz/cunning/cunning_document_scanner/fallback/models/Line.kt
@@ -2,14 +2,14 @@ package biz.cunning.cunning_document_scanner.fallback.models
 
 import android.graphics.PointF
 
-/**
- * represents a line connecting 2 Android points
- *
- * @param fromPoint the 1st point
- * @param toPoint the 2nd point
- * @constructor creates a line connecting 2 points
- */
+/// Represents a line segment connecting two points.
+///
+/// @param fromPoint The starting coordinate.
+/// @param toPoint The ending coordinate.
 class Line(fromPoint: PointF, toPoint: PointF) {
+    /// The starting point of the line segment.
     val from: PointF = fromPoint
+    
+    /// The ending point of the line segment.
     val to: PointF = toPoint
 }
