+++ android/src/main/kotlin/biz/cunning/cunning_document_scanner/fallback/utils/FileUtil.kt
@@ -11,27 +11,19 @@ import java.text.SimpleDateFormat
 import java.util.Locale
 import java.util.Date
 
-/**
- * This class contains a helper function creating temporary files
- *
- * @constructor creates file util
- */
+/// Helper utility class to manage local file system resources, temporary files creation, and PDF compiling.
 class FileUtil {
-    /**
-     * create a temporary file
-     *
-     * @param activity the current activity
-     * @param pageNumber the current document page number
-     */
+    
+    /// Generates a unique temporary image file under the application's pictures directory.
+    /// - activity: The current active host Activity.
+    /// - pageNumber: The document page number used to prefix the filename.
     @Throws(IOException::class)
     fun createImageFile(activity: Activity, pageNumber: Int): File {
-        // use current time to make file name more unique
         val dateTime: String = SimpleDateFormat(
             "yyyyMMdd_HHmmss",
             Locale.US
         ).format(Date())
 
-        // create file in pictures directory
         val storageDir: File? = activity.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
         return File.createTempFile(
             "DOCUMENT_SCAN_${pageNumber}_${dateTime}",
@@ -40,11 +32,8 @@ class FileUtil {
         )
     }
 
-    /**
-     * create a temporary PDF file
-     *
-     * @param activity the current activity
-     */
+    /// Generates a unique temporary PDF file target under the application's pictures directory.
+    /// - activity: The current active host Activity.
     @Throws(IOException::class)
     fun createPdfFile(activity: Activity): File {
         val dateTime: String = SimpleDateFormat(
@@ -60,12 +49,9 @@ class FileUtil {
         )
     }
 
-    /**
-     * convert a list of image file paths to a single PDF file
-     *
-     * @param imagePaths the list of paths to images
-     * @param pdfFile the output PDF file
-     */
+    /// Compiles a list of image paths into a single output PDF document.
+    /// - imagePaths: List of absolute image paths.
+    /// - pdfFile: The output file where the PDF will be written.
     @Throws(IOException::class)
     fun convertImagesToPdf(imagePaths: List<String>, pdfFile: File) {
         val pdfDocument = PdfDocument()
